Added save_sgf and load_sgf to leela Moved winrate_transformer function to utils
