+++ sgftools/utils.py
@@ -1,3 +1,6 @@
+import math
+import sys
+
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
@@ -102,3 +105,78 @@ def join_list_into_str(list_to_join, separator):
     :return: string
     """
     return separator.join(map(str, list_to_join))
+
+
+def winrate_transformer(stdev, verbosity):
+    """
+    Stretch winrate out the middle range and squashe the extreme ranges
+    :return: winrate
+    """
+    def variance(k):
+        """
+        Variance of the distribution
+        :param k: 0 <= k <= 1
+        """
+        k = float(k)
+        return 0.25 - (k ** 2 + 2 * k + 1) / (2 * k ** 2 + 5 * k + 3) / 2
+
+    def find_k(lower, upper):
+        """
+        Binary search to find the appropriate k
+        """
+        while True:
+            mid = 0.5 * (lower + upper)
+            if mid == lower or mid == upper or lower >= upper:
+                return mid
+            var = variance(mid)
+            if var < stdev * stdev:
+                upper = mid
+            else:
+                lower = mid
+
+    if stdev * stdev <= 1e-10:
+        raise ValueError("Stdev too small, please choose a more reasonable value")
+
+    # Repeated doubling to find an upper bound big enough
+    upper = 1
+    while variance(upper) > stdev * stdev:
+        upper = upper * 2
+
+    k = find_k(0, upper)
+
+    if verbosity > 2:
+        print("Using k = %f, stdev = %f" % (k, math.sqrt(variance(k))), file=sys.stderr)
+
+    def unnormpdf(x):
+        """
+        Unnormalize probability density function
+        """
+        if x <= 0 or x >= 1:
+            return 0
+        a = math.log(x)
+        b = math.log(1 - x)
+        logprob = a * k + b * k
+        # Constant scaling so we don't overflow floats with crazy values
+        logprob = logprob - 2 * k * math.log(0.5)
+        return math.exp(logprob)
+
+    # Precompute a big array to approximate the CDF
+    n = 100000
+    lookup = [unnormpdf(float(x) / float(n)) for x in range(n + 1)]
+    cum = 0
+
+    for i in range(n + 1):
+        cum += lookup[i]
+        lookup[i] = cum
+
+    for i in range(n + 1):
+        lookup[i] = lookup[i] / lookup[n]
+
+    def cdf(x):
+        i = int(math.floor(x * n))
+        if i >= n or i < 0:
+            return x
+        excess = x * n - i
+        return lookup[i] + excess * (lookup[i + 1] - lookup[i])
+
+    return lambda x: cdf(x)
